# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sutradhar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/items": {
      "post": {
        "tags": [
          "provenance"
        ],
        "summary": "Register Item",
        "description": "Register an item. Item, event, and outbox row commit together or not at all.",
        "operationId": "register_item_api_v1_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "List Items",
        "description": "Keyset page over items, newest first.",
        "operationId": "list_items_api_v1_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Slug"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ItemStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "registered_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered By"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/items/{item_id}/tag": {
      "post": {
        "tags": [
          "qr"
        ],
        "summary": "Issue a tag code for an item",
        "description": "Generate a code, bind it to the item, and record a ``TAG_ISSUED`` event.\n\n409 when the item is already tagged, with the existing code in the error\ndetails: that is a fact to read, not a failure to retry.",
        "operationId": "issue_tag_api_v1_items__item_id__tag_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/items/{item_id}/tag/qr": {
      "get": {
        "tags": [
          "qr"
        ],
        "summary": "Render an item's tag as a QR image",
        "description": "PNG by default, SVG on request. Error correction Q, four-module quiet zone.\n\n``size`` is a pixel edge for the PNG and the default ``width``/``height`` on\nthe SVG, which also carries a ``viewBox`` so it scales to any label.",
        "operationId": "render_qr_api_v1_items__item_id__tag_qr_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(png|svg)$",
              "default": "png",
              "title": "Format"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 512,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/png": {},
              "image/svg+xml": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tags/bulk": {
      "post": {
        "tags": [
          "qr",
          "admin"
        ],
        "summary": "Issue tags for many items at once",
        "description": "Tag a batch, reporting each item separately.\n\nAn already-tagged item or one whose anchoring failed is reported and\nstepped over. Rejecting the whole batch for one bad row would make an\noperator find and remove it by hand before any of the good ones printed.",
        "operationId": "bulk_issue_api_v1_admin_tags_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkTagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkTagRequest": {
        "properties": {
          "item_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Item Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_ids"
        ],
        "title": "BulkTagRequest",
        "description": "Items to tag in one pass."
      },
      "BulkTagResponse": {
        "properties": {
          "requested": {
            "type": "integer",
            "title": "Requested"
          },
          "issued": {
            "type": "integer",
            "title": "Issued"
          },
          "already_tagged": {
            "type": "integer",
            "title": "Already Tagged"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkTagResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requested",
          "issued",
          "already_tagged",
          "failed",
          "results"
        ],
        "title": "BulkTagResponse",
        "description": "Per-item results and their totals. Partial success is a success."
      },
      "BulkTagResult": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "tag_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Code"
          },
          "display_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Code"
          },
          "payload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Url"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "outcome"
        ],
        "title": "BulkTagResult",
        "description": "What happened to one item in the batch."
      },
...
      "DisputeStatus": {
        "type": "string",
        "enum": [
          "NONE",
          "DISPUTED"
        ],
        "title": "DisputeStatus",
        "description": "Whether an item's provenance is contested."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pagination"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "ItemListResponse",
        "description": "Collection envelope with a keyset cursor, per app.core.pagination."
      },
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "FAILED"
        ],
        "title": "ItemStatus",
        "description": "Anchoring state of an item's hash. PENDING until the chain confirms."
      },
      "ItemSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category_schema_version": {
            "type": "integer",
            "title": "Category Schema Version"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "registered_by": {
            "type": "string",
            "format": "uuid",
            "title": "Registered By"
          },
          "quantity": {
            "type": "string",
            "title": "Quantity"
          },
          "quantity_unit": {
            "type": "string",
            "title": "Quantity Unit"
          },
          "item_hash": {
            "type": "string",
            "title": "Item Hash"
          },
          "tag_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Code"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "dispute_status": {
            "$ref": "#/components/schemas/DisputeStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "category_schema_version",
          "parent_id",
          "registered_by",
          "quantity",
          "quantity_unit",
          "item_hash",
          "tag_code",
          "status",
          "dispute_status",
          "created_at"
        ],
        "title": "ItemSummary",
        "description": "An item in a listing."
      },
...
      "RegisterItemRequest": {
        "properties": {
          "category_slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Category Slug"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Quantity"
          },
          "quantity_unit": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Quantity Unit"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "category_slug",
          "attributes",
          "quantity",
          "quantity_unit"
        ],
        "title": "RegisterItemRequest",
        "description": "Register a new item, or a root of a new provenance tree."
      },
...
      "TagResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "tag_code": {
            "type": "string",
            "title": "Tag Code"
          },
          "display_code": {
            "type": "string",
            "title": "Display Code"
          },
          "payload_url": {
            "type": "string",
            "title": "Payload Url"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "tag_code",
          "display_code",
          "payload_url"
        ],
        "title": "TagResponse",
        "description": "One issued tag, in both stored and printable form."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}